+++ apps/recnet-api/src/utils/auth/auth.guard.ts
@@ -1,7 +1,7 @@
 import {
   CanActivate,
   ExecutionContext,
-  Injectable,
+  mixin,
   UnauthorizedException,
 } from "@nestjs/common";
 import { Request } from "express";
@@ -10,29 +10,40 @@ import { getPublicKey } from "@recnet/recnet-jwt";
 
 import { VerifyJwtFunction } from "./auth.type";
 
-@Injectable()
-export class AuthGuard implements CanActivate {
-  constructor(private readonly verifyJwt: VerifyJwtFunction) {}
+export const AuthGuard = (
+  verifyJwt: VerifyJwtFunction,
+  isOptional: boolean
+) => {
+  class AuthGuardMixin implements CanActivate {
+    async canActivate(context: ExecutionContext) {
+      const request = context.switchToHttp().getRequest();
+      const token = this.extractTokenFromHeader(request);
 
-  async canActivate(context: ExecutionContext) {
-    const request = context.switchToHttp().getRequest();
-    const token = this.extractTokenFromHeader(request);
+      if (!token) {
+        if (isOptional) {
+          return true;
+        }
+        throw new UnauthorizedException();
+      }
 
-    if (!token) {
-      throw new UnauthorizedException();
+      try {
+        const publicKey = await getPublicKey(token);
+        const payload = verifyJwt(token, publicKey);
+        request.user = payload;
+      } catch (error) {
+        if (isOptional) {
+          return true;
+        }
+        throw new UnauthorizedException();
+      }
+      return true;
     }
-    try {
-      const publicKey = await getPublicKey(token);
-      const payload = this.verifyJwt(token, publicKey);
-      request.user = payload;
-    } catch (error) {
-      throw new UnauthorizedException();
-    }
-    return true;
-  }
 
-  private extractTokenFromHeader(request: Request): string | undefined {
-    const [type, token] = request.headers.authorization?.split(" ") ?? [];
-    return type === "Bearer" ? token : undefined;
+    private extractTokenFromHeader(request: Request): string | undefined {
+      const [type, token] = request.headers.authorization?.split(" ") ?? [];
+      return type === "Bearer" ? token : undefined;
+    }
   }
-}
+  const guard = mixin(AuthGuardMixin);
+  return guard;
+};