+++ apps/recnet/src/components/RecCard.tsx
@@ -1,5 +1,5 @@
-import { CalendarIcon } from "@radix-ui/react-icons";
-import { Flex, Text } from "@radix-ui/themes";
+import { CalendarIcon, CheckCircledIcon } from "@radix-ui/react-icons";
+import { Flex, Text, Tooltip } from "@radix-ui/themes";
 import { ChevronRight } from "lucide-react";
 import Link from "next/link";
 
@@ -17,6 +17,7 @@ import { LinkCopyButton } from "./LinkCopyButton";
 import { SelfRecBadge } from "./SelfRecBadge";
 
 import { getSharableLink } from "../utils/getSharableRecLink";
+import { RecReactionsList } from "../app/rec/[id]/RecReactionsList";
 
 export function RecCardSkeleton() {
   return (
@@ -105,6 +106,14 @@ export function RecCard(props: { recs: Rec[]; showDate?: boolean }) {
             <Flex className="items-center gap-x-2 text-gray-10">
               <CalendarIcon width={16} height={16} />
               <Text size="2">{`${!rec.article.month ? "" : `${numToMonth[rec.article.month]}, `}${rec.article.year}`}</Text>
+              {rec.article.isVerified ? (
+                <Tooltip content="This article comes from trusted sources.">
+                  <div className="flex items-center gap-x-1 cursor-pointer mx-2">
+                    <CheckCircledIcon />
+                    <Text size="2">Verified</Text>
+                  </div>
+                </Tooltip>
+              ) : null}
             </Flex>
             <Flex className="items-center gap-x-1 text-accent-11">
               <Text size="1">Read</Text>{" "}
@@ -152,6 +161,9 @@ export function RecCard(props: { recs: Rec[]; showDate?: boolean }) {
                   </Text>
                 </Flex>
               </Link>
+              <div className="mt-2">
+                <RecReactionsList id={rec.id} />
+              </div>
             </Flex>
           </Flex>
         );